feat: refactor venue profile types and move to utils
